requirements.txt
streamlit
selenium
google-generativeai
python-dotenv
pandas